feature: melhora estrutura de diretórios e implementa restante da regra de negócio

+++ src/main/java/org/tech_solutions/application/user/dto/UserDataDTO.java
@@ -1,9 +1,10 @@
 package org.tech_solutions.application.user.dto;
 
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 public record UserDataDTO(
-        Long id,
+        BigInteger id,
         String name,
         String email,
         LocalDateTime created_at,
